% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_table_right.R
\name{add_table_right}
\alias{add_table_right}
\title{Adds aggregation column to table}
\usage{
add_table_right(ftable, ffun = sum, colname = "Result", fcols = NA)
}
\arguments{
\item{ftable}{data.frame or data.table}

\item{ffun}{aggregation function such as sum, mean, etc.
Default value is sum.}

\item{colname}{string. Default value is 'Result'.}

\item{fcols}{vector. Column names to be included in the
aggregation (only if numeric). Default value is NA, i.e. all
columns are included.}
}
\value{
data.frame or data.table
}
\description{
Adds aggregation column to table
}
\examples{
\dontrun{add_table_right(dt)}
}
