% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ecbar.R
\name{plot_ecbar}
\alias{plot_ecbar}
\title{Interactive Bar Plot}
\usage{
plot_ecbar(dp = NULL)
}
\arguments{
\item{dp}{List of parameters

\itemize{
\item tab: table to be plotted, already aggregated
\item xval = 'x': column name with x-values
\item yval = 'y': column name with y-values
\item group = 'group': column name for grouping (not needed)
\item color = 'color': column name with color-values
\item margin_left = '10\%': chart margin left
\item margin_right = '12': chart margin right
\item text_format = '': text format is '' (standard), 'percent', or 'euro'
\item legend = FALSE: show (TRUE/FALSE) the legend
\item title: chart title
}}
}
\value{
an echarts4r bar plot
}
\description{
Interactive Bar Plot
}
\examples{
\dontrun{plot_ecbar(dp)}
\dontrun{
dp = list(NULL)
dp$tab = tab
dp$xval = 'x'
dp$yval = 'y'
dp$group = 'group'
dp$color = 'color'
dp$text_format = ''
dp$margin_left = '10\%'
dp$margin_right = '12'
dp$title = 'Chart'
dp$legend = FALSE
fplot = plot_ecpar(dp)
fplot
}

}
