% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ecdot.R
\name{plot_ecdot}
\alias{plot_ecdot}
\title{Interactive Dot or Cleveland Plot}
\usage{
plot_ecdot(dp = NULL)
}
\arguments{
\item{dp}{List of parameters

\itemize{
\item tab: table to be plotted, already aggregated
\item xval = 'x': column name with x-values, categorical or integer values
\item yval = 'y': column name with y-values, numerical values
\item group = 'group': column name with group values, categorical values
\item color = 'color': column name with colors
\item pointsize = 8: symbol size
\item linesize = 2: line width of line plot
\item show_line = TRUE: show horizontal lines (TRUE/FALSE)
\item line_color = '#A9A9A9': color of horizontal lines
\item title: chart title
\item text_format = '': text format is '' (standard), 'percent', or 'euro'
\item show_label = TRUE: show text labels (TRUE/FALSE)
\item label_color = '#A9A9A9': color of text labels
\item label_size = 10: size of text labels
\item label_position = 'out': position of text labels ('out', 'left', 'right', 'inside', 'top', 'bottom')
\item x_label_color = '#606060': color of x-axis text labels
\item x_label_size = 15: size of x-axis text labels
\item y_label_color = '#606060': color of y-axis text labels
\item y_label_size = 10: size of y-axis text labels
\item xaxis_title = '': name of x-axis title (string), set NA to remove it
\item show_xaxis_line = FALSE: show (vertical) x-axis line
\item show_yaxis = FALSE: show (horizontal) y-axis
\item yaxis_title = yval: name of y-axis title (string), set NA to remove it
\item axis_title_color = '#606060': color of axes titles
\item axis_title_size = 15: size of axes titles
\item show_xgrid = FALSE: show x-grid (TRUE/FALSE)
\item show_ygrid = FALSE: show y-grid (TRUE/FALSE)
\item margin_left = '10\%': chart margin left
\item margin_right = '12': chart margin right
}}
}
\value{
an echarts4r dot or cleveland plot
}
\description{
Interactive Dot or Cleveland Plot
}
\examples{
\dontrun{plot_ecdot(dp)}
\dontrun{
dp = list(NULL)
dp$tab = tab
dp$xval = 'x'
dp$yval = 'y'
dp$group = 'group'
dp$color = 'color'
dp$pointsize = 8
dp$linesize = 2
dp$show_line = TRUE
dp$line_color = '#A9A9A9'
dp$title = NULL
dp$text_format = ''
dp$show_label = FALSE
dp$label_color = '#A9A9A9'
dp$label_size = 10
dp$label_position = 'out'
dp$x_label_color = '#606060'
dp$x_label_size = 15
dp$y_label_color = '#606060'
dp$y_label_size = 10
dp$xaxis_title = NULL
dp$show_xaxis_line = FALSE
dp$show_yaxis = FALSE
dp$yaxis_title = NULL
dp$axis_title_color = '#606060'
dp$axis_title_size = 15
dp$show_xgrid = FALSE
dp$show_ygrid = FALSE
dp$margin_left = '10\%'
dp$margin_right = '12'
fplot = plot_dot(dp)
fplot
}

}
