% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_heatmap.R
\name{plot_echeatmap}
\alias{plot_echeatmap}
\title{Interactive Heatmap Plot}
\usage{
plot_echeatmap(dp = NULL)
}
\arguments{
\item{dp}{List of parameters

\itemize{
\item tab: table to be plotted, already aggregated
\item xval = 'x': column name with x-values, categorical or integer values
\item yval = 'y': column name with y-values, categorical or integer values
\item zval = 'z': column name with z-values, numerical values
\item color = 'c('#fef8cb', '#ba8d7e'): color vector defining color palette
\item title = NULL: chart title
\item title_left = margin_left: left position of title
\item title_top  = 'auto': top position of title
\item title_size = 18: text size of title
\item title_color = x_label_color: text color of title
\item text_format = '': text format is '' (standard), 'percent', or 'euro'
\item show_label = TRUE: show text labels (TRUE/FALSE)
\item label_color = '#606060': color of text labels
\item label_size = 10: size of text labels
\item x_label_color = '#606060': color of x-axis text labels
\item x_label_size = 15: size of x-axis text labels
\item y_label_color = '#606060': color of y-axis text labels
\item y_label_size = 15: size of y-axis text labels
\item xaxis_title = xval: name of x-axis title (string), set NA to remove it
\item yaxis_title = yval: name of y-axis title (string), set NA to remove it
\item axis_title_color = '#606060': color of axes titles
\item axis_title_size = 15: size of axes titles
\item show_xgrid = FALSE: show x-grid (TRUE/FALSE)
\item show_ygrid = FALSE: show y-grid (TRUE/FALSE)
\item margin_left = '10\%': chart margin left
\item margin_right = '10\%': chart margin right
\item margin_top = 60: chart margin top
\item margin_bottom = 60: chart margin bottom
\item chart_width = NULL: chart width (css units)
\item chart_height = NULL: chart height (css units)
}}
}
\value{
an echarts4r heatmap plot
}
\description{
Interactive Heatmap Plot
}
\examples{
\dontrun{plot_echeatmap(dp)}
\dontrun{
dp = list(NULL)
dp$tab = tab
dp$xval = 'x'
dp$yval = 'y'
dp$zval = 'z'
dp$color = 'c('#fef8cb', '#ba8d7e')
dp$text_format = ''
dp$show_label = TRUE
dp$label_color = '#606060'
dp$label_size = 10
dp$x_label_color = '#606060'
dp$x_label_size = 15
dp$y_label_color = '#606060'
dp$y_label_size = 15
dp$xaxis_title = xval
dp$yaxis_title = yval
dp$axis_title_color = '#606060'
dp$axis_title_size = 15
dp$show_xgrid = FALSE
dp$show_ygrid = FALSE
dp$margin_left = '10\%'
dp$margin_right = '10\%'
dp$margin_top = 60
dp$margin_bottom = 60
dp$title = 'Chart'
dp$chart_width = NULL
dp$chart_height = NULL
fplot = plot_echeatmap(dp)
fplot
}

}
