% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ecscatter.R
\name{plot_ecscatter}
\alias{plot_ecscatter}
\title{Interactive Bar Plot}
\usage{
plot_ecscatter(dp = NULL)
}
\arguments{
\item{dp}{List of parameters

\itemize{
\item tab: table to be plotted, already aggregated
\item xval = 'x': column name with x-values
\item yval = 'y': column name with y-values
\item group = 'group': column name for grouping (not needed)
\item color = NULL: color vector
\item pointsize = 8: symbol size
\item linesize = 2: line width of line plot
\item title: chart title
\item text_format = '': text format is '' (standard), 'percent', or 'euro'
\item show_legend = FALSE: show legend (TRUE/FALSE)
\item show_label = FALSE: show text labels (TRUE/FALSE)
\item label_color = '#A9A9A9': color of text labels
\item label_size = 10: size of text labels
\item ax_label_color = '#606060': color of axes text labels
\item ax_label_size = 10: size of axes text labels
\item xaxis_title = xval: name of x-axis title (string), set NA to remove it
\item yaxis_title = yval: name of y-axis title (string), set NA to remove it
\item axis_title_color = '#606060': color of axes titles
\item axis_title_size = 15: size of axes titles
\item show_xgrid = FALSE: show x-grid (TRUE/FALSE)
\item show_ygrid = FALSE: show y-grid (TRUE/FALSE)
\item margin_left = '10\%': chart margin left
\item margin_right = '12': chart margin right
}}
}
\value{
an echarts4r bar plot
}
\description{
Interactive Bar Plot
}
\examples{
\dontrun{plot_ecbar(dp)}
\dontrun{
dp = list(NULL)
dp$tab = tab
dp$xval = 'x'
dp$yval = 'y'
dp$group = 'group'
dp$color = c('#b4943e', '#777acd', '#60a862', '#c45ca2', '#cb5a4c')
dp$pointsize = 8
dp$linesize = 2
dp$title = NULL
dp$text_format = ''
dp$show_legend = FALSE
dp$show_label = FALSE
dp$label_color = '#A9A9A9'
dp$label_size = 10
dp$ax_label_color = '#606060'
dp$ax_label_size = 10
dp$xaxis_title = NULL
dp$yaxis_title = NULL
dp$axis_title_color = '#606060'
dp$axis_title_size = 15
dp$show_xgrid = FALSE
dp$show_ygrid = FALSE
dp$margin_left = '10\%'
dp$margin_right = '12'
fplot = plot_ecscatter(dp)
fplot
}

}
