% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ggwaffle.R
\name{plot_ggwaffle}
\alias{plot_ggwaffle}
\title{Waffle Plot}
\usage{
plot_ggwaffle(dp)
}
\arguments{
\item{dp}{List of parameters

\itemize{
\item tab: table to be plotted, aggregated values
\item key_label = '': column name with plot labels
\item key_family = '': column name with font family
\item key_frequ = '': column name with frequencies
\item key_color = '': column name with colors
\item key_group = '': column name with groups
\item key_facet = '': column name with facet groups
\item label = '\\u25A0': labels to be plotted, only if key_label is not defined
\item family = NA: font family for labels, only if key_family is not defined
\item frequ = 1: frequency, only if key_frequ is not defined
\item color = '#bb873c': vector with colors, only if key_color is not defined

\item cols = 5: number of columns in waffle chart
\item size = 10: size of plotted labels
\item orientation = 'vertical': orientation of waffel plot pillars (vertical or horizontal)
\item align = 'bottom': alignment of waffle plot pillars (top, bottom, center).
E.g. centered facet pillars
\item facet_ncol = 5: maximum number of facet columns

\item legend = 'none': legend position (top, right, left, bottom, none, label)
'label' uses labels as legend symbols, works only in some situation with
position always right
\item legend_symbol_size = 3: size of legend symbols
\item legend_text_size = 10: #'
}}
}
\value{
a ggplot waffle chart
}
\description{
Waffle Plot
}
\examples{
\dontrun{plot_ecbar(dp)}
\dontrun{
dp = list(NULL)
dp$tab = tab
dp$xval = 'x'
dp$yval = 'y'
dp$group = 'group'
dp$color = 'color'
dp$text_format = ''
dp$label_color = '#A9A9A9'
dp$label_size = 10
dp$x_label_color = '#606060'
dp$x_label_size = 15
dp$margin_left = '10\%'
dp$margin_right = '12'
dp$title = 'Chart'
dp$chart_width = NULL
dp$chart_height = NULL
dp$legend = FALSE
dp$legend_left = 'auto'
dp$legend_right = 'auto'
dp$legend_top = 'auto'
dp$legend_bottom = 'auto'
dp$legend_orient = 'horizontal'
fplot = plot_ecbar(dp)
fplot
}

}
