% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_table.R
\name{explore_table}
\alias{explore_table}
\title{Summary statistics and plots of the table columns.}
\usage{
explore_table(
  ftable,
  ftarget = NA,
  fassign_classes = TRUE,
  fadd_tables = TRUE,
  fadd_plots = FALSE,
  fprm_na = FALSE,
  fpmax_numlevels = 0,
  fpmax_faclevels = 50
)
}
\arguments{
\item{ftable}{data.frame or data.table}

\item{ftarget}{string: columname of ftable including the target.
Optional. Use ftarget = NA to explore the table without target column.}

\item{fassign_classes}{TRUE or FALSE. TRUE assigns a new class to each column.}

\item{fadd_tables}{TRUE or FALSE. Add or not add a table.}

\item{fadd_plots}{TRUE or FALSE. Add or not add a plot.}

\item{fprm_na}{TRUE or FALSE. Remove or not remove NAs before plotting.}

\item{fpmax_numlevels}{Integer. If number of numerics is smaller, then plot
bar plot instead of line plot.}

\item{fpmax_faclevels}{Integer. Maximum number of bars to be plotted for each segment.}
}
\value{
list with several elements:
\emph{summary}: the summary statistics.
\emph{plot$name$count}: ggplot of absolute frequencies of property \emph{name}.
\emph{plot$name$rel}: ggplot of relative frequencies of property \emph{name}.
}
\description{
Summary statistics and plots of the table columns.
}
\examples{
\dontrun{explore_table(dt)}
}
