% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ecbar.R
\name{plot_ecbar}
\alias{plot_ecbar}
\title{Interactive Bar Plot}
\usage{
plot_ecbar(dp = NULL)
}
\arguments{
\item{dp}{List of parameters

\itemize{
\item tab: table to be plotted, already aggregated
\item xval = 'x': column name with x-values
\item yval = 'y': column name with y-values
\item group = 'group': column name for grouping (not needed)
\item color = 'color': column name with color-values
\item label_color = '#A9A9A9': color of text labels
\item x_label_color = '#606060': color of x-axis text labels
\item label_size = 10: size of text labels
\item x_label_size = 15: size of x-axis text labels
\item margin_left = '10\%': chart margin left
\item margin_right = '12': chart margin right
\item text_format = '': text format is '' (standard), 'percent', or 'euro'
\item chart_width = NULL: chart width (css units)
\item chart_height = NULL: chart height (css units)
\item legend = FALSE: show (TRUE/FALSE) the legend
\item legend_left = 'auto': legend position left
\item legend_right = 'auto': legend position right
\item legend_top = 'auto': legend position top
\item legend_bottom = 'auto': legend position bottom
\item legend_orient = 'horizontal': legend orientation (horizontal or vertical)
\item title: chart title
}}
}
\value{
an echarts4r bar plot
}
\description{
Interactive Bar Plot
}
\examples{
\dontrun{plot_ecbar(dp)}
\dontrun{
dp = list(NULL)
dp$tab = tab
dp$xval = 'x'
dp$yval = 'y'
dp$group = 'group'
dp$color = 'color'
dp$text_format = ''
dp$label_color = '#A9A9A9'
dp$label_size = 10
dp$x_label_color = '#606060'
dp$x_label_size = 15
dp$margin_left = '10\%'
dp$margin_right = '12'
dp$title = 'Chart'
dp$chart_width = NULL
dp$chart_height = NULL
dp$legend = FALSE
dp$legend_left = 'auto'
dp$legend_right = 'auto'
dp$legend_top = 'auto'
dp$legend_bottom = 'auto'
dp$legend_orient = 'horizontal'
fplot = plot_ecbar(dp)
fplot
}

}
