% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ecscatter.R
\name{plot_ecscatter}
\alias{plot_ecscatter}
\title{Interactive Line and Scatter Plot}
\usage{
plot_ecscatter(dp = NULL)
}
\arguments{
\item{dp}{List of parameters

\itemize{
\item tab: table to be plotted, already aggregated
\item xval = 'x': column name with x-values
\item yval = 'y': column name with y-values
\item group = 'group': column name for grouping (not needed)
\item color = NULL: color vector
\item pointsize = 8: symbol size
\item linesize = 2: line width of line plot
\item title: chart title
\item text_format = '': text format is '' (standard), 'percent', or 'euro'
\item show_label = FALSE: show text labels (TRUE/FALSE)
\item label_color = '#A9A9A9': color of text labels
\item label_size = 10: size of text labels
\item ax_label_color = '#606060': color of axes text labels
\item ax_label_size = 10: size of axes text labels
\item xaxis_title = xval: name of x-axis title (string), set NA to remove it
\item yaxis_title = yval: name of y-axis title (string), set NA to remove it
\item axis_title_color = '#606060': color of axes titles
\item axis_title_size = 15: size of axes titles
\item show_xgrid = FALSE: show x-grid (TRUE/FALSE)
\item show_ygrid = FALSE: show y-grid (TRUE/FALSE)
\item margin_left = '10\%': chart margin left
\item margin_right = '10\%': chart margin right
\item margin_top = 60: chart margin top
\item margin_bottom = 60: chart margin bottom
\item legend = FALSE: show (TRUE/FALSE) the legend
\item legend_left = 'auto': legend position left
\item legend_right = 'auto': legend position right
\item legend_top = 'auto': legend position top
\item legend_bottom = 'auto': legend position bottom
\item legend_orient = 'horizontal': legend orientation (horizontal or vertical)
\item chart_width = NULL: chart width (css units)
\item chart_height = NULL: chart height (css units)
\item x_min = 'dataMin': min. value on x-axis
\item x_max = NULL: max. value on x-axis
\item y_min = NULL: min. value on y-axis
\item y_max = NULL: max. value on y-axis
}}
}
\value{
an echarts4r line and scatter plot
}
\description{
Interactive Line and Scatter Plot
}
\examples{
\dontrun{plot_ecbar(dp)}
\dontrun{
dp = list(NULL)
dp$tab = tab
dp$xval = 'x'
dp$yval = 'y'
dp$group = 'group'
dp$color = c('#b4943e', '#777acd', '#60a862', '#c45ca2', '#cb5a4c')
dp$pointsize = 8
dp$linesize = 2
dp$title = NULL
dp$text_format = ''
dp$show_label = FALSE
dp$label_color = '#A9A9A9'
dp$label_size = 10
dp$ax_label_color = '#606060'
dp$ax_label_size = 10
dp$xaxis_title = NULL
dp$yaxis_title = NULL
dp$axis_title_color = '#606060'
dp$axis_title_size = 15
dp$show_xgrid = FALSE
dp$show_ygrid = FALSE
dp$margin_left = '10\%'
dp$margin_right = '10\%'
dp$margin_top = 60
dp$margin_bottom = 60
dp$legend = FALSE
dp$legend_left = 'auto'
dp$legend_right = 'auto'
dp$legend_top = 'auto'
dp$legend_bottom = 'auto'
dp$legend_orient = 'horizontal'
dp$chart_width = NULL
dp$chart_height = NULL
fplot = plot_ecscatter(dp)
fplot
}

}
