% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ggwaffle.R
\name{plot_ggwaffle}
\alias{plot_ggwaffle}
\title{Waffle Plot}
\usage{
plot_ggwaffle(dp)
}
\arguments{
\item{dp}{List of parameters

\itemize{
\item tab: table to be plotted, aggregated values
\item key_label = '': column name with plot labels
\item key_family = '': column name with font family
\item key_frequ = '': column name with frequencies
\item key_color = '': column name with colors
\item key_group = '': column name with groups
\item key_facet = '': column name with facet groups
\item label = '\\u25A0': labels to be plotted, only if key_label is not defined
\item family = NA: font family for labels, only if key_family is not defined
\item frequ = 1: frequency, only if key_frequ is not defined
\item color = '#bb873c': vector with colors, only if key_color is not defined

\item cols = 5: number of columns in waffle chart
\item size = 10: size of plotted labels
\item orientation = 'vertical': orientation of waffel plot pillars (vertical or horizontal)
\item align = 'bottom': alignment of waffle plot pillars (top, bottom, center).
E.g. centered facet pillars
\item facet_ncol = 5: maximum number of facet columns

\item legend = 'none': legend position (top, right, left, bottom, none, label)
'label' uses labels as legend symbols, works only in some situations with
position always right
\item legend_symbol_size = 3: size of legend symbols
\item legend_text_size = 15: size of legend text
\item label_legend_size = 10: size of label-legend symbols and text
\item label_legend_color = '#A9A9A9': color of label-legend text

\item facet_margin = 2: (-2 if legend = 'label'), inner margin between neighboring
facets.
\item chart_xy_margin = 1: (3 if legend = 'label' to give enough space),
extra xy-limit for each plot (increases x-axis limits)
\item margin_bottom = 1: to increase y-axis at bottom side
\item margin_top = 1: to increase y-axis at top side
\item outermargin_left = 0: chart margin left
\item outermargin_right = 1: chart margin right

\item facet_text_size = 25: size of facet title
\item facet_text_color = '#707070': color of facet title
\item facet_title_position = 'bottom': position of facet title (top, left, bottom, right)

\item font = '': font family of chart

}}
}
\value{
a ggplot waffle chart
}
\description{
Waffle Plot
}
\examples{
\dontrun{plot_ggwaffel(dp)}
\dontrun{
dp = list(NULL)
dp$tab = tab
fplot = plot_ggwaffel(dp)
fplot
}

}
